+++ src/hooks/useGames.ts
@@ -1,0 +1,32 @@
+import { useEffect, useState } from "react";
+import { CanceledError } from "axios";
+import apiClient from "../service/api-client";
+
+interface FetchGamesResponse {
+  count: number;
+  results: Game[];
+}
+
+interface Game {
+  id: number;
+  name: string;
+}
+const useGames = () => {
+  const [games, setGames] = useState<Game[]>([]);
+  const [error, setError] = useState("");
+  const controller = new AbortController();
+
+  useEffect(() => {
+    apiClient
+      .get("/games")
+      .then((res) => setGames(res.data.results))
+      .catch((err) => {
+        if (err instanceof CanceledError) return;
+        setError(err.message);
+      });
+    return () => controller.abort();
+  }, []);
+  return { games, error };
+};
+
+export default useGames;
